+++ MetalLiqViajes_Bll_Data/dal/RutasOrigenDestinoVehTrailerDataProvider.cs
@@ -57,8 +57,8 @@ namespace LiqViajes_Bll_Data
 		/// <param name="TipoTrailerCodigo"></param>
 		/// <param name="DescripcionTipoTrailer"></param>
 		/// <param name="Favorita"></param>
-		/// <returns>int that contents the Codigo value</returns>
-		public int Create(int Codigo, int? RutasOrigenDestinoCodigo, string Origen, string Destino, string GrupoOrigen, string GrupoDestino, int? TipoVehiculoCodigo, string TipoVehiculo, int? TipoTrailerCodigo, string DescripcionTipoTrailer, bool? Favorita,string module, Sinapsys.Datos.SQL datosTransaccion=null)
+		/// <returns>long that contents the Codigo value</returns>
+		public long Create(long Codigo, long? RutasOrigenDestinoCodigo, string Origen, string Destino, string GrupoOrigen, string GrupoDestino, int? TipoVehiculoCodigo, string TipoVehiculo, int? TipoTrailerCodigo, string DescripcionTipoTrailer, bool? Favorita,string module, Sinapsys.Datos.SQL datosTransaccion=null)
 		{
 			try 
 			{
@@ -124,7 +124,7 @@ namespace LiqViajes_Bll_Data
 					paramlist.AddWithValue("@Favorita",Favorita);
 				}
 				// Execute the query and return the new identity value
-				int returnValue = Convert.ToInt32(LocalDataProvider.EjecutarProcedimiento("dbo.RutasOrigenDestinoVehTrailerCreate", paramlist, disconnect, out nullExit, DataProvider.TiempoEspera).Rows[0][0]);
+				long returnValue = Convert.ToInt32(LocalDataProvider.EjecutarProcedimiento("dbo.RutasOrigenDestinoVehTrailerCreate", paramlist, disconnect, out nullExit, DataProvider.TiempoEspera).Rows[0][0]);
 
 				return returnValue;
 			}
@@ -149,7 +149,7 @@ namespace LiqViajes_Bll_Data
 		/// <param name="TipoTrailerCodigo"></param>
 		/// <param name="DescripcionTipoTrailer"></param>
 		/// <param name="Favorita"></param>
-		public void Update(int Codigo, int? RutasOrigenDestinoCodigo, string Origen, string Destino, string GrupoOrigen, string GrupoDestino, int? TipoVehiculoCodigo, string TipoVehiculo, int? TipoTrailerCodigo, string DescripcionTipoTrailer, bool? Favorita,string module, Sinapsys.Datos.SQL datosTransaccion=null)
+		public void Update(long Codigo, long? RutasOrigenDestinoCodigo, string Origen, string Destino, string GrupoOrigen, string GrupoDestino, int? TipoVehiculoCodigo, string TipoVehiculo, int? TipoTrailerCodigo, string DescripcionTipoTrailer, bool? Favorita,string module, Sinapsys.Datos.SQL datosTransaccion=null)
 		{
 			try 
 			{
@@ -226,7 +226,7 @@ namespace LiqViajes_Bll_Data
 		/// Deletes one record from RutasOrigenDestinoVehTrailer by passing all key fields
 		/// </summary>
 		/// <param name="Codigo"></param>
-		public void Delete(int Codigo,string module, Sinapsys.Datos.SQL datosTransaccion=null)
+		public void Delete(long Codigo,string module, Sinapsys.Datos.SQL datosTransaccion=null)
 		{
 			try 
 			{
@@ -264,7 +264,7 @@ namespace LiqViajes_Bll_Data
 		/// </summary>
 		/// <param name="Codigo"></param>
 		/// <returns>A DataTable object containing the data</returns>
-		public DataTable Get(int Codigo)
+		public DataTable Get(long Codigo)
 		{
 			try 
 			{
@@ -331,7 +331,7 @@ namespace LiqViajes_Bll_Data
 		/// </summary>
 		/// <param name="RutasOrigenDestinoCodigo"></param>
 		/// <returns>A DataTable object containing all records data</returns>
-		public DataTable GetBy_RutasOrigenDestinoCodigo(int RutasOrigenDestinoCodigo)
+		public DataTable GetBy_RutasOrigenDestinoCodigo(long RutasOrigenDestinoCodigo)
 		{
 			try 
 			{
