+++ MetalLiqViajes_Bll_Data/dal/RutasOrigenTipoTrailerDataProvider.cs
@@ -51,8 +51,8 @@ namespace LiqViajes_Bll_Data
 		/// <param name="Origen"></param>
 		/// <param name="TipoTrailerCodigo"></param>
 		/// <param name="DescripcionTrailer"></param>
-		/// <returns>int that contents the Codigo value</returns>
-		public int Create(int Codigo, int? RutasOrigenCodigo, string Origen, int? TipoTrailerCodigo, string DescripcionTrailer,string module, Sinapsys.Datos.SQL datosTransaccion=null)
+		/// <returns>long that contents the Codigo value</returns>
+		public long Create(long Codigo, long? RutasOrigenCodigo, string Origen, int? TipoTrailerCodigo, string DescripcionTrailer,string module, Sinapsys.Datos.SQL datosTransaccion=null)
 		{
 			try 
 			{
@@ -94,7 +94,7 @@ namespace LiqViajes_Bll_Data
 					paramlist.AddWithValue("@DescripcionTrailer",DescripcionTrailer);
 				}
 				// Execute the query and return the new identity value
-				int returnValue = Convert.ToInt32(LocalDataProvider.EjecutarProcedimiento("dbo.RutasOrigenTipoTrailerCreate", paramlist, disconnect, out nullExit, DataProvider.TiempoEspera).Rows[0][0]);
+				long returnValue = Convert.ToInt32(LocalDataProvider.EjecutarProcedimiento("dbo.RutasOrigenTipoTrailerCreate", paramlist, disconnect, out nullExit, DataProvider.TiempoEspera).Rows[0][0]);
 
 				return returnValue;
 			}
@@ -113,7 +113,7 @@ namespace LiqViajes_Bll_Data
 		/// <param name="Origen"></param>
 		/// <param name="TipoTrailerCodigo"></param>
 		/// <param name="DescripcionTrailer"></param>
-		public void Update(int Codigo, int? RutasOrigenCodigo, string Origen, int? TipoTrailerCodigo, string DescripcionTrailer,string module, Sinapsys.Datos.SQL datosTransaccion=null)
+		public void Update(long Codigo, long? RutasOrigenCodigo, string Origen, int? TipoTrailerCodigo, string DescripcionTrailer,string module, Sinapsys.Datos.SQL datosTransaccion=null)
 		{
 			try 
 			{
@@ -166,7 +166,7 @@ namespace LiqViajes_Bll_Data
 		/// Deletes one record from RutasOrigenTipoTrailer by passing all key fields
 		/// </summary>
 		/// <param name="Codigo"></param>
-		public void Delete(int Codigo,string module, Sinapsys.Datos.SQL datosTransaccion=null)
+		public void Delete(long Codigo,string module, Sinapsys.Datos.SQL datosTransaccion=null)
 		{
 			try 
 			{
@@ -204,7 +204,7 @@ namespace LiqViajes_Bll_Data
 		/// </summary>
 		/// <param name="Codigo"></param>
 		/// <returns>A DataTable object containing the data</returns>
-		public DataTable Get(int Codigo)
+		public DataTable Get(long Codigo)
 		{
 			try 
 			{
